fix(mcp): pass structural-gate by eliminating import-cycle regressions
The sentrux structural gate failed on cycle + quality + complexity
regressions. Resolved without weakening the baseline:

- extract makeHandle/freshConnect/listToolDefs into a leaf handle.ts so
  create-mcp-tools and rehydrate no longer form a cycle
- move MCPTransportKind to a leaf transport-types.ts, breaking the
  types <-> cache-types cycle
- import Tool/tool from @openrouter/agent's deep subpaths
  (/tool, /tool-types) instead of the barrel, so mcp doesn't absorb the
  agent lib SCC
- drop the unused ExecuteToolsResult interface in agent's tool-types.ts,
  which removes the tool-types -> model-result edge and breaks agent's
  pre-existing lib cycle (this is what mcp was being dragged into)
- collapse tryCacheHit's option forwarding into a data-driven helper to
  drop it back under the max-cyclomatic-complexity threshold

Gate now: cycles 1->1, quality 0.51->0.52, no degradation.
