+++ packages/mcp/src/rehydrate.ts
@@ -3,8 +3,8 @@ import type { MCPCacheStore } from './cache/cache-store.js';
 import { defaultCacheKey } from './cache/cache-store.js';
 import type { SerializedMCPServer } from './cache/cache-types.js';
 import { isSerializedMCPServer } from './cache/cache-types.js';
-import { freshConnect, makeHandle } from './create-mcp-tools.js';
 import { MCPCacheError } from './errors.js';
+import { freshConnect, makeHandle } from './handle.js';
 import { connect } from './mcp-connection.js';
 import type { UnconvertibleSchemaMode } from './schema/json-schema-to-zod.js';
 import type { McpToolDef } from './tool-wrapper.js';
