+++ packages/mcp/src/transport-types.ts
@@ -1,0 +1,2 @@
+/** Non-stdio MCP transports this package supports. */
+export type MCPTransportKind = 'streamableHttp' | 'sse';
